python-app/pyproject.toml
[project]
name = "python-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.28.1",
    "logfire[httpx,system-metrics]>=4.7.0",
    "loguru>=0.7.3",
    "pydantic>=2.11.9",
    "python-dotenv>=1.1.1",
]
